+++ netbox_aws_vpc_plugin/api/serializers.py
@@ -1,6 +1,8 @@
 from rest_framework import serializers
 
 from ipam.api.serializers import PrefixSerializer
+from dcim.api.serializers import RegionSerializer
+from tenancy.api.serializers import TenantSerializer
 from netbox.api.serializers import NetBoxModelSerializer
 from ..models import AWSVPC, AWSSubnet, AWSAccount
 from .nested_serializers import NestedAWSAccountSerializer, NestedAWSVPCSerializer
@@ -10,6 +12,7 @@ class AWSVPCSerializer(NetBoxModelSerializer):
     url = serializers.HyperlinkedIdentityField(view_name="plugins-api:netbox_aws_vpc_plugin-api:awsvpc-detail")
     vpc_cidr = PrefixSerializer(required=False, allow_null=True, default=None, nested=True)
     owner_account = NestedAWSAccountSerializer()
+    region = RegionSerializer(required=False, allow_null=True, default=None, nested=True)
 
     class Meta:
         model = AWSVPC
@@ -22,6 +25,7 @@ class AWSVPCSerializer(NetBoxModelSerializer):
             "arn",
             "vpc_cidr",
             "owner_account",
+            "region",
             "tags",
             "custom_fields",
             "created",
@@ -34,6 +38,7 @@ class AWSSubnetSerializer(NetBoxModelSerializer):
     subnet_cidr = PrefixSerializer(required=False, allow_null=True, default=None, nested=True)
     vpc = NestedAWSVPCSerializer()
     owner_account = NestedAWSAccountSerializer()
+    region = RegionSerializer(required=False, allow_null=True, default=None, nested=True)
 
     class Meta:
         model = AWSSubnet
@@ -47,6 +52,7 @@ class AWSSubnetSerializer(NetBoxModelSerializer):
             "subnet_cidr",
             "vpc",
             "owner_account",
+            "region",
             "tags",
             "custom_fields",
             "created",
@@ -56,6 +62,7 @@ class AWSSubnetSerializer(NetBoxModelSerializer):
 
 class AWSAccountSerializer(NetBoxModelSerializer):
     url = serializers.HyperlinkedIdentityField(view_name="plugins-api:netbox_aws_vpc_plugin-api:awsaccount-detail")
+    tenant = TenantSerializer(required=False, allow_null=True, default=None, nested=True)
 
     class Meta:
         model = AWSAccount
@@ -66,6 +73,7 @@ class AWSAccountSerializer(NetBoxModelSerializer):
             "account_id",
             "name",
             "arn",
+            "tenant",
             "description",
             "tags",
             "custom_fields",