+++ netbox_aws_vpc_plugin/models/aws_account.py
@@ -8,17 +8,14 @@ from netbox.models import NetBoxModel
 
 
 class AWSAccount(NetBoxModel):
-    account_id = models.CharField(max_length=12, unique=True)
-    arn = models.CharField(
-        max_length=2000, blank=True, verbose_name="The Amazon Resource Name (ARN) of the account root."
-    )
+    account_id = models.CharField(max_length=12, unique=True, verbose_name="Account ID")
+    arn = models.CharField(max_length=2000, blank=True, verbose_name="ARN")
     name = models.CharField(
         max_length=50,
         blank=True,
     )
     description = models.CharField(max_length=500, blank=True)
-    # TODO: Netbox Tenant
-    # TODO: AWS Org
+    tenant = models.ForeignKey(to="tenancy.Tenant", on_delete=models.PROTECT, blank=True, null=True)
     # TODO: Status
     comments = models.TextField(blank=True)